+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Updated Discord integration to use discord.com instead of discordapp.com
 - Add "Failure Keyword" filtering for inbound emails (#396)
 - Add support for multiple, comma-separated keywords (#396)
+- New integration: phone calls (#403)
 
 ### Bug Fixes
 - Removing Pager Team integration, project appears to be discontinued